requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
